+++ app/src/main/java/de/baumann/browser/activity/BrowserActivity.java
@@ -262,7 +262,7 @@ public class BrowserActivity extends AppCompatActivity implements BrowserControl
                 .putInt("restart_changed", 0)
                 .putBoolean("pdf_create", false).putBoolean("redirect", sp.getBoolean("sp_youTube_switch", false) || sp.getBoolean("sp_twitter_switch", false) || sp.getBoolean("sp_instagram_switch", false))
                 .putString("profile", sp.getString("profile_toStart", "profileStandard"))
-                .putString("dialog_neverAsk", "no").apply();
+                .putString("dialog_neverAsk", "no").putString("dialog_neverAskBackGround", "no").apply();
 
         switch (Objects.requireNonNull(sp.getString("start_tab", "3"))) {
             case "3":
@@ -2431,7 +2431,7 @@ public class BrowserActivity extends AppCompatActivity implements BrowserControl
             addAlbum(null, data, true, false, "", null);
             getIntent().setAction("");
             hideOverview();
-            BrowserUnit.openInBackground(activity, intent, data);
+            BrowserUnit.openInBackground(activity, ninjaWebView);
         }
     }